Separate registry per python version
